모듈화 수정
