+++ signin.html
@@ -54,7 +54,7 @@
       </div>
     </div>
   </div>
-  <script src="signin.js"></script>
+  <script type="module" src="signin.js"></script>
 </body>
 
 </html>
