+++ signup.html
@@ -65,7 +65,7 @@
       </div>
     </div>
   </div>
-  <script src="signup.js"></script>
+  <script type="module" src="signup.js"></script>
 </body>
 
 </html>
