+++ signup.js
@@ -1,3 +1,5 @@
+import { validateEmail, validatePassword, togglePasswordVisibility } from './sign_moduel.js';
+
 document.addEventListener('DOMContentLoaded', function () {
   var emailInput = document.querySelector('#email');
   var passwordInput = document.querySelector('#password');
@@ -7,35 +9,21 @@ document.addEventListener('DOMContentLoaded', function () {
   var confirmPasswordError = document.querySelector('#confirm-password-error');
   var eyeButtons = document.querySelectorAll('.eye-button');
 
-  function validateEmail() {
-    var email = emailInput.value;
-    if (email.length === 0) {
-      emailError.textContent = '이메일을 입력해주세요.';
-      emailError.style.display = 'block';
-      emailInput.classList.add('input-error');
-    } else if (!email.includes('@')) {
-      emailError.textContent = '올바른 이메일 주소가 아닙니다.';
-      emailError.style.display = 'block';
-      emailInput.classList.add('input-error');
-    } else {
-      emailError.style.display = 'none';
-      emailInput.classList.remove('input-error');
-    }
-  }
+  eyeButtons.forEach(function (button) {
+    button.addEventListener('click', function () {
+      togglePasswordVisibility(this);
+    });
+  });
 
-  function validatePassword() {
-    var password = passwordInput.value;
-    if (password.length === 0) {
-      passwordError.textContent = '비밀번호를 입력해주세요.';
-      passwordError.style.display = 'block';
-      passwordInput.classList.add('input-error');
-    } else {
-      passwordError.style.display = 'none';
-      passwordInput.classList.remove('input-error');
-    }
-  }
+  emailInput.addEventListener('focusout', function () {
+    validateEmail(emailInput, emailError);
+  });
 
-  function validateConfirmPassword() {
+  passwordInput.addEventListener('focusout', function () {
+    validatePassword(passwordInput, passwordError);
+  });
+
+  confirmPasswordInput.addEventListener('focusout', function () {
     if (passwordInput.value !== confirmPasswordInput.value) {
       confirmPasswordError.textContent = '비밀번호가 일치하지 않습니다.';
       confirmPasswordError.style.display = 'block';
@@ -44,31 +32,12 @@ document.addEventListener('DOMContentLoaded', function () {
       confirmPasswordError.style.display = 'none';
       confirmPasswordInput.classList.remove('input-error');
     }
-  }
-
-  eyeButtons.forEach(function (button) {
-    button.addEventListener('click', function () {
-      var passwordField = button.previousElementSibling;
-      var eyeIcon = button.querySelector('.eye-icon');
-      if (passwordField.type === 'password') {
-        passwordField.type = 'text';
-        eyeIcon.src = './images/eye.png';
-      } else {
-        passwordField.type = 'password';
-        eyeIcon.src = './images/eye-off.svg';
-      }
-    });
   });
-
-  emailInput.addEventListener('focusout', validateEmail);
-  passwordInput.addEventListener('focusout', validatePassword);
-  confirmPasswordInput.addEventListener('focusout', validateConfirmPassword);
 
   var signupButton = document.querySelector('.cta');
   signupButton.addEventListener('click', function (event) {
     event.preventDefault();
-    validateEmail();
-    validatePassword();
-    validateConfirmPassword();
+    validateEmail(emailInput, emailError);
+    validatePassword(passwordInput, passwordError);
   });
 });
